LPD-91502 Rename ViewDashboard to Dashboards
