+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/index.ts
@@ -49,7 +49,7 @@ export {default as CategorizationToolbar} from './main_view/categorization/Categ
 export {default as EditCategoryPage} from './main_view/categorization/categories/EditCategoryPage';
 export {default as ViewTags} from './main_view/categorization/tags/ViewTags';
 export {default as EditVocabulary} from './main_view/categorization/vocabularies/EditVocabulary';
-export {default as ViewDashboard} from './main_view/dashboard/ViewDashboard';
+export {default as Dashboards} from './main_view/dashboard/Dashboards';
 
 export {default as EditFolder} from './main_view/folders/EditFolder';
 export {default as QuickActions} from './main_view/home/QuickActions';
